Extract the pair session's tool half into PairToolExecutor
